+++ settings.gradle
@@ -19,3 +19,4 @@ include 'whitebird-commons:whitebird-jpa-spring-boot-starter'
 include 'whitebird-commons:whitebird-redis-spring-boot-starer'
 include 'whitebird-commons:whitebird-web-spring-boot-starter'
 include 'whitebird-commons:whitebird-swagger-spring-boot-starter'
+include 'whitebird-commons:whitebird-oauth2-spring-boot-starter'
