+++ whitebird-auth-service/src/main/java/top/idwangmo/whitebird/authservice/config/Oauth2WebConfig.java
@@ -1,11 +1,12 @@
 package top.idwangmo.whitebird.authservice.config;
 
 import org.springframework.context.annotation.Configuration;
-import top.idwangmo.whitebird.commoncore.config.DefaultWebMvcConfig;
+import top.idwangmo.whitebird.oauth2springbootstarter.config.DefaultOauth2WebConfig;
 
 /**
  * @author idwangmo
  */
 @Configuration
-public class WebMvcConfig extends DefaultWebMvcConfig {
+public class Oauth2WebConfig extends DefaultOauth2WebConfig {
+
 }
