+++ whitebird-commons/whitebird-core-spring-boot-starter/build.gradle
@@ -10,20 +10,17 @@ description = 'whitebird-common-core-starter'
 
 dependencies {
     implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
-    implementation 'org.springframework.cloud:spring-cloud-starter-oauth2'
     compileOnly 'org.projectlombok:lombok'
-    developmentOnly 'org.springframework.boot:spring-boot-devtools'
     annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
     annotationProcessor 'org.projectlombok:lombok'
     testImplementation('org.springframework.boot:spring-boot-starter-test') {
         exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
     }
 
-    // 第三方包
-    implementation 'cn.hutool:hutool-all:5.0.3'
-    compileOnly 'javax.servlet:javax.servlet-api:4.0.1'
-    implementation 'com.google.code.gson:gson:2.8.6'
-
+    // 第三包
+    implementation "cn.hutool:hutool-all:${hutool}"
+    implementation "com.google.code.gson:gson:${gson}"
+    implementation "javax.servlet:javax.servlet-api:${javaxServlet}"
     implementation "io.github.resilience4j:resilience4j-ratelimiter:${resilience4jVersion}"
 }
