--- whitebird-commons/whitebird-core-spring-boot-starter/src/main/java/top/idwangmo/whitebird/commoncore/mapper/WhitebirdUserMapper.java
@@ -1,4 +1,0 @@
-package top.idwangmo.whitebird.commoncore.mapper;
-
-public interface WhitebirdUserMapper {
-}
