+++ whitebird-commons/whitebird-oauth2-spring-boot-starter/build.gradle
@@ -1,0 +1,37 @@
+plugins {
+	id 'org.springframework.boot' version "${springBootVersion}"
+	id 'io.spring.dependency-management' version "${springDependencyManagement}"
+	id 'java'
+	id 'maven'
+	id 'maven-publish'
+}
+
+dependencies {
+	implementation 'org.springframework.cloud:spring-cloud-starter-oauth2'
+	compileOnly 'org.projectlombok:lombok'
+	annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
+	annotationProcessor 'org.projectlombok:lombok'
+	testImplementation('org.springframework.boot:spring-boot-starter-test') {
+		exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
+	}
+
+	implementation project(':whitebird-commons').project('whitebird-core-spring-boot-starter')
+
+	implementation "com.google.code.gson:gson:${gson}"
+	implementation "cn.hutool:hutool-all:${hutool}"
+	compileOnly "javax.servlet:javax.servlet-api:${javaxServlet}"
+}
+
+dependencyManagement {
+	imports {
+		mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
+	}
+}
+
+jar {
+	enabled = true
+}
+
+bootJar {
+	enabled = false
+}
