+++ whitebird-commons/whitebird-oauth2-spring-boot-starter/src/main/java/top/idwangmo/whitebird/oauth2springbootstarter/annotation/CurrentUser.java
@@ -1,4 +1,4 @@
-package top.idwangmo.whitebird.commoncore.annotation;
+package top.idwangmo.whitebird.oauth2springbootstarter.annotation;
 
 import java.lang.annotation.*;
