+++ whitebird-commons/whitebird-oauth2-spring-boot-starter/src/main/java/top/idwangmo/whitebird/oauth2springbootstarter/config/WhitebirdAuthExceptionConfigure.java
@@ -1,8 +1,8 @@
-package top.idwangmo.whitebird.commoncore.config;
+package top.idwangmo.whitebird.oauth2springbootstarter.config;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
-import top.idwangmo.whitebird.commoncore.handler.WhitebirdAccessDeniedHandler;
+import top.idwangmo.whitebird.oauth2springbootstarter.handler.WhitebirdAccessDeniedHandler;
 
 /**
  * auth 异常配置.
